Push code into dedicated GitDiff component
