+++ packages/cli-kit/package.json
@@ -123,6 +123,7 @@
     "fs-extra": "11.1.0",
     "fuzzy": "0.1.3",
     "get-port-please": "3.0.1",
+    "git-diff": "^2.0.6",
     "gradient-string": "2.0.2",
     "graphql": "16.4.0",
     "graphql-request": "5.2.0",
@@ -158,6 +159,7 @@
     "@types/commondir": "^1.0.0",
     "@types/cross-zip": "^4.0.0",
     "@types/diff": "^5.0.3",
+    "@types/git-diff": "^2.0.0",
     "@types/fs-extra": "9.0.13",
     "@types/gradient-string": "^1.1.2",
     "@types/lodash": "4.14.195",
